feat: Add write-bin subcommand
